Fixed compile warnings

+++ src/SFML/Network/Packet.cpp
@@ -405,7 +405,7 @@ Packet& Packet::operator <<(double data)
 Packet& Packet::operator <<(const char* data)
 {
     // First insert string length
-    Uint32 length = std::strlen(data);
+    Uint32 length = static_cast<Uint32>(std::strlen(data));
     *this << length;
 
     // Then insert characters
@@ -434,7 +434,7 @@ Packet& Packet::operator <<(const std::string& data)
 Packet& Packet::operator <<(const wchar_t* data)
 {
     // First insert string length
-    Uint32 length = std::wcslen(data);
+    Uint32 length = static_cast<Uint32>(std::wcslen(data));
     *this << length;
 
     // Then insert characters
